Pass rootWindow to all sagas
